public-api/requirements.txt
aiohttp==3.8.1
diff_match_patch
websocket-client
requests
cerberus